Calculate Brier score

A function for calculating the Brier score for a survival model.

brier_score(y_true = NULL, risk = NULL, surv = NULL, times = NULL)

loss_brier_score(y_true = NULL, risk = NULL, surv = NULL, times = NULL)

Arguments

y_true

a survival::Surv object containing the times and statuses of observations for which the metric will be evaluated

risk

ignored, left for compatibility with other metrics

surv

a matrix containing the predicted survival functions for the considered observations, each row represents a single observation, whereas each column one time point

times

a vector of time points at which the survival function was evaluated

Value

numeric from 0 to 1, lower scores are better (Brier score of 0.25 represents a model which returns always returns 0.5 as the predicted survival function)

Details

Brier score is used to evaluate the performance of a survival model, based on the squared distance between the predicted survival function and the actual event time, weighted to account for censored observations.

References

[1] Brier, Glenn W. "Verification of forecasts expressed in terms of probability." Monthly Weather Review 78.1 (1950): 1-3.

[2] Graf, Erika, et al. "Assessment and comparison of prognostic classification schemes for survival data." Statistics in Medicine 18.17‐18 (1999): 2529-2545.

See also

Examples

library(survival)
library(survex)

cph <- coxph(Surv(time, status) ~ ., data = veteran, model = TRUE, x = TRUE, y = TRUE)
cph_exp <- explain(cph)
#> Preparation of a new explainer is initiated 
#>   -> model label       :  coxph (  default  ) 
#>   -> data              :  137  rows  6  cols (  extracted from the model  ) 
#>   -> target variable   :  137  values ( 128 events and 9 censored , censoring rate = 0.066 ) (  extracted from the model  ) 
#>   -> times             :  49 unique time points , min = 1 , mean = 110.6971 , median = 80 , max = 562.52 
#>   -> times             :  (  generated from y as 50 time points being consecutive quantiles (0.00, 0.02, ..., 0.98)  ) 
#>   -> predict function  :  predict.coxph with type = 'risk' will be used (  default  ) 
#>   -> predict survival function  :  predictSurvProb.coxph will be used (  default  ) 
#>   -> predict cumulative hazard function  :  -log(predict_survival_function) will be used (  default  ) 
#>   -> model_info        :  package survival , ver. 3.5.3 , task survival (  default  ) 
#>   A new explainer has been created!  

y <- cph_exp$y
times <- cph_exp$times
surv <- cph_exp$predict_survival_function(cph, cph_exp$data, times)

brier_score(y, surv = surv, times = times)
#>  [1] 0.01450469 0.02131799 0.05085721 0.07237285 0.08296505 0.10034947
#>  [7] 0.11860639 0.12115921 0.12534671 0.12899805 0.12894844 0.13452743
#> [13] 0.13853090 0.14201661 0.15028547 0.15745581 0.15887866 0.16678636
#> [19] 0.16054595 0.15386626 0.15595910 0.15824493 0.16763382 0.16343522
#> [25] 0.15672279 0.15781375 0.15707059 0.15188724 0.15110733 0.15151055
#> [31] 0.15325489 0.16278271 0.16469163 0.16725319 0.16644868 0.16572450
#> [37] 0.16073207 0.15352345 0.14642744 0.14359704 0.13666323 0.12885560
#> [43] 0.11816123 0.10765794 0.09889997 0.07760702 0.06668832 0.04565290
#> [49] 0.02744170
loss_brier_score(y, surv = surv, times = times)
#>  [1] 0.01450469 0.02131799 0.05085721 0.07237285 0.08296505 0.10034947
#>  [7] 0.11860639 0.12115921 0.12534671 0.12899805 0.12894844 0.13452743
#> [13] 0.13853090 0.14201661 0.15028547 0.15745581 0.15887866 0.16678636
#> [19] 0.16054595 0.15386626 0.15595910 0.15824493 0.16763382 0.16343522
#> [25] 0.15672279 0.15781375 0.15707059 0.15188724 0.15110733 0.15151055
#> [31] 0.15325489 0.16278271 0.16469163 0.16725319 0.16644868 0.16572450
#> [37] 0.16073207 0.15352345 0.14642744 0.14359704 0.13666323 0.12885560
#> [43] 0.11816123 0.10765794 0.09889997 0.07760702 0.06668832 0.04565290
#> [49] 0.02744170