Adapt mlr3proba measures for use with survex

This function allows for usage of standardized measures from the mlr3proba package with survex.

loss_adapt_mlr3proba(measure, reverse = FALSE, ...)

Arguments

measure

a MeasureSurv object from the mlr3proba package, the object to adapt

reverse

boolean, FALSE by default, whether the metric should be reversed in order to be treated as loss (for permutational variable importance we need functions with lower values indicating better performance). If TRUE, the new metric value will be (1 - metric_value)

...

other parameters, currently ignored

Value

a function with standardized parameters (y_true, risk, surv, times) that can be used to calculate loss

if(FALSE) measure <- msr("surv.calib_beta") mlr_measure <- loss_adapt_mlr3proba(measure)