Helper functions for model_performance.R

Helper functions for model_performance.R

surv_model_performance(
  explainer,
  ...,
  times = NULL,
  type = "metrics",
  metrics = NULL
)

Arguments

explainer

an explainer object - model preprocessed by the explain() function

...

other parameters, currently ignored

times

a numeric vector, time points at which ROC curves are calculated if type == "roc" or at which metrics are calculated if type == "metrics". Note: if type=="roc" this parameter is obligatory

type

character, either "metrics" which calculates performance metrics or "roc" which calculates ROC curves at specific time points

metrics

a named vector containing the metrics to be calculated. The values should be standardized loss functions. The functions can be supplied manually but has to have these named parameters (y_true, risk, surv, times), where y_true represents the survival::Surv object with observed times and statuses, risk is the risk score calculated by the model, and surv is the survival function for each observation evaluated at times.

Value

Either a list when type == "metrics" or a data.frame if type == "roc"