Helper functions for predict_parts.R

Helper functions for predict_parts.R

Usage

surv_shap(
  explainer,
  new_observation,
  ...,
  y_true = NULL,
  calculation_method = "kernel",
  aggregation_method = "integral",
  path = "average",
  B = 25,
  exact = FALSE
)

Arguments

explainer

an explainer object - model preprocessed by the explain() function

new_observation

a new observation for which predictions need to be explained

...

additional parameters, passed to internal functions

y_true

a two element numeric vector or matrix of one row and two columns, the first element being the true observed time and the second the status of the observation, used for plotting

calculation_method

a character, only "kernel" is implemented for now.

aggregation_method

a character, either "mean_absolute" or "integral", "max_absolute", "sum_of_squares"

path

ignored, placeholder the not implemented "sampling" method

B

ignored, placeholder the not implemented "sampling" method

exact

ignored, placeholder the not implemented "sampling" method

Value

A list, containing the calculated SurvSHAP(t) results in the result field

References

[1] Krzyziński, Mateusz, et al. "SurvSHAP(t): Time-dependent explanations of machine learning survival models." arXiv preprint arXiv:2208.11080 (2022).